feat: Remove unnecessary utilities and rename utility files
