+++ client/src/utils/index.js
@@ -1,8 +1,8 @@
-export * as postDate from './postDate';
+export * as postDate from './postDateUtils';
 export * as propTypeSchema from './propTypeSchema';
 export * as reduxUtils from './reduxUtils';
 export * as validationSchema from './validationSchema';
 export * as dataFilteringUtils from './dataFilteringUtils';
 export * as formHandler from './handler/formHandler';
-export * as path from './path';
+export * as path from './pathUtils';
 export { default as URI } from './apiUtils';
